+++ packages/terra-form-checkbox/examples/test-examples/checkbox/MultipleCheckboxes.example.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+import Checkbox from '../../../lib/Checkbox';
+
+const checkbox = () => (
+  <div>
+    <Checkbox id="first" labelText="First Checkbox" defaultChecked />
+    <Checkbox id="second" labelText="Second Checkbox" />
+    <Checkbox id="third" labelText="Third Checkbox" />
+  </div>
+);
+
+export default checkbox;