+++ packages/terra-form-checkbox/tests/nightwatch/DefaultCheckbox.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import Checkbox from '../../lib/Checkbox';
-
-const checkbox = () => (<Checkbox id="default" labelText="Default Checkbox" />);
-
-export default checkbox;