+++ packages/terra-form-checkbox/tests/nightwatch/DisabledCheckbox.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-import Checkbox from '../../lib/Checkbox';
-
-const checkbox = () => (<Checkbox id="disabled" labelText="Disabled Checkbox" disabled />);
-
-export default checkbox;